Marlin-compatible gcode
Changed order of gcode so laser power is now prefaced with the 'laser on' command and is on the preceding line to the move command.

Moved 'Fxx' speed command to a new line prefaced by G91.

Have not tested vertical or diagonal raster order yet.  May have screwed up diagonal.